Fixes #15 - Reads from Postgres were not updating 'active' in thing db layer.
This problem came from our never having implemented the reads from the
database of the 'active' column. The column was properly written to the
database but things did not have a valid _active property, this was
being set to the same value of the date by default. So when the cache
did not contain the thing, then the read from the db would return the
wrong value for _active. If the cache did contain the thing after thing
creation, then the correct value would be there until the cache was
cleared.