Re: #94 Notify button and form
There's now a notify button for each post, with a form to notify a list
of users. It posts to POST_notify in api.py. It only works in a space
or in the comment screen, the namepicker doesn't yet work for it, and
it isn't properly validated yet. But it is a place to create some
actual notifications, by just typing in one or more usernames.