Re: #178 #88 Space name picker
Replaced the built-in space picker with the same look and code as the
other name pickers. Currently showing spaces the user is able to write
to. Only active for adding new items right now. To be added in more
places later. Space header images should be added too, when they're
ready.