Re #88: Namepicker Safari problem
I was loading the list of users & spaces asynchronously, which causes
some problems in some browsers, as the code that activates the
namepicker(s) might load before that list has been picked up. Turning
off the asynchronous loading for now, which shouldn't be much of a
problem, as it only needs to be picked up occasionally (once per hour
currently) and the list is small.